app/requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.0
django-allauth==0.41.0
django-appconf==1.0.3
django-celery==3.3.1
django-celery-email==3.0.0
django-celery-results==1.2.0
django-markdownx==3.0.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==1.5.0
kombu==4.6.7
Markdown==3.2
oauthlib==3.1.0
Pillow==6.2.0
psycopg2-binary==2.8.3
python3-openid==3.1.0
pytz==2019.3
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
zipp==2.2.0